feat: integrate layout
