+++ .prettierrc.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  semi: false,
+  singleQuote: true,
+  trailingComma: 'all',
+  printWidth: 100,
+  proseWrap: 'always',
+  plugins: [require('prettier-plugin-tailwindcss')],
+}
