--- .prettierrc.json
@@ -1,7 +1,0 @@
-{
-  "semi": false,
-  "singleQuote": true,
-  "trailingComma": "all",
-  "printWidth": 100,
-  "proseWrap": "always"
-}
