+++ package.json
@@ -62,6 +62,7 @@
     "lint-staged": "=13.2.1",
     "postcss": "^8.4.24",
     "prettier": "^2.8.7",
+    "prettier-plugin-tailwindcss": "^0.3.0",
     "sass": "^1.62.1",
     "tailwindcss": "^3.3.2",
     "typescript": "^5.0.4",
