+++ src/index.css
@@ -1,3 +1,15 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+/* Apply this to top level containers to fix their layout shift when scrollbar appears */
+@media (min-width: 1024px /* breakpoint for desktop layout */ ) {
+  .top-level-container {
+    padding-left: calc(100vw - 100%);
+  }
+}
+
+/* The warning is a bit distracting over time */
+.firebase-emulator-warning {
+  opacity: 0.25;
+}
