+++ src/layout/RootLayout.tsx
@@ -2,8 +2,8 @@ import styles from './RootLayout.module.css'
 import { Outlet } from 'react-router-dom'
 import Logo from './logo/Logo.tsx'
 import Menu from './Menu.tsx'
-import ErrorBoundary from '../../core/react/display/ErrorBoundary.tsx'
-import PageViewLogger from '../../core/firebase/tracking/PageViewLogger.tsx'
+import ErrorBoundary from '../core/react/display/ErrorBoundary.tsx'
+import PageViewLogger from '../core/firebase/tracking/PageViewLogger.tsx'
 import { PropsWithChildren } from 'react'
 
 const RootLayout = ({ children }: PropsWithChildren): JSX.Element => {
