+++ src/layout/logo/Logo.tsx
@@ -1,5 +1,5 @@
 import styles from './Logo.module.css'
-import { ReactComponent as LogoSvg } from '../../../assets/logo.svg'
+import { ReactComponent as LogoSvg } from '../../assets/logo.svg'
 
 const Logo = (): JSX.Element => {
   return (
