+++ src/pages/Pages.tsx
@@ -2,17 +2,22 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import ErrorPage from './error/ErrorPage.tsx'
 import Practice from './practice/Practice.tsx'
 import { Authentication } from '../core/firebase/auth/Authentication.tsx'
-import RootLayout2 from './root/RootLayout2.tsx'
+import { Layout } from '../layout/Layout.tsx'
+import { Components } from './components/Components.tsx'
 
 const router = createBrowserRouter([
   {
     path: '/',
-    element: <RootLayout2 />,
+    element: <Layout />,
     errorElement: <ErrorPage />,
     children: [
       {
         path: '/',
         element: <Authentication />,
+      },
+      {
+        path: '/components',
+        element: <Components />,
       },
       {
         path: 'practice',
